+++ libft/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: mgruson <mgruson@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/09 13:06:30 by mgruson           #+#    #+#             */
-/*   Updated: 2022/05/18 13:37:32 by mgruson          ###   ########.fr       */
+/*   Updated: 2022/05/18 15:41:34 by mgruson          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void *ft_memcpy(void *dest, const void *src, size_t n)
 {
-    int i;
+    size_t i;
 
     i = 0;
-    while(i != n)
+    while (i != n)
     {
         *(char *)dest++ = *(char *)src++;
         i++;
