+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: mgruson <mgruson@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/09 10:35:38 by mgruson           #+#    #+#             */
-/*   Updated: 2022/05/18 13:38:46 by mgruson          ###   ########.fr       */
+/*   Updated: 2022/05/18 15:47:46 by mgruson          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int i;
+    size_t i;
 
     i = 0;
     while(s1[i] != '\0' && s2[i] != '\0' && s1[i] == s2[i] && i < n - 1)
